Move files into LAUNCH POSITION
