--- javascript_nodejs/10.prompt-validations/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-npm-debug.log
-.env
