+++ samples/javascript_nodejs/10.prompt-validations/.env
@@ -1,0 +1,1 @@
+botFilePath=prompt-validations-bot.bot
