+++ samples/javascript_nodejs/10.prompt-validations/.eslintrc.js
