+++ samples/javascript_nodejs/10.prompt-validations/.gitignore
@@ -1,0 +1,2 @@
+node_modules/
+npm-debug.log
