+++ samples/javascript_nodejs/10.prompt-validations/prompt-validations-bot.bot
@@ -1,11 +1,11 @@
 {
     "name": "prompt-validations-bot",
-    "secretKey": "",
+    "padlock": "",
     "services": [
         {
             "type": "endpoint",
-            "name": "prompt-validations-bot",
-            "id": "http://localhost:3978/api/messages",
+            "name": "development",
+            "id": "1",
             "appId": "",
             "appPassword": "",
             "endpoint": "http://localhost:3978/api/messages"
