+++ samples/javascript_nodejs/10.prompt-validations/prompts/colorPrompt/index.js
@@ -5,7 +5,7 @@ const { ChoicePrompt } = require('botbuilder-dialogs');
 
 // This is a custom choice prompt that will emit an error if the user
 // types an invalid choice.
-module.exports = class ColorPrompt extends ChoicePrompt {
+module.exports.ColorPrompt = class ColorPrompt extends ChoicePrompt {
     constructor(dialogId) {
         super(dialogId, async (prompt) => {
             if (!prompt.recognized.succeeded) {
