+++ samples/javascript_nodejs/10.prompt-validations/prompts/namePrompt/index.js
@@ -4,7 +4,7 @@
 const { TextPrompt } = require('botbuilder-dialogs');
 
 // This is a custom TextPrompt that requires the input to be between 1 and 50 characters in length.
-module.exports = class NamePrompt extends TextPrompt {
+module.exports.NamePrompt = class NamePrompt extends TextPrompt {
     constructor(dialogId) {
         super(dialogId, async (prompt) => {
             if (!prompt.recognized.succeeded) {
